Move command handler method test env. in a sub-package
