+++ server/src/test/java/io/spine/server/command/model/CommandHandlerMethodTest.java
@@ -30,23 +30,23 @@ import io.spine.core.CommandEnvelope;
 import io.spine.core.Event;
 import io.spine.server.aggregate.Aggregate;
 import io.spine.server.command.AbstractCommandHandler;
-import io.spine.server.command.given.CommandHandlerMethodTestEnv.HandlerReturnsEmpty;
-import io.spine.server.command.given.CommandHandlerMethodTestEnv.HandlerReturnsEmptyList;
-import io.spine.server.command.given.CommandHandlerMethodTestEnv.InvalidHandlerNoAnnotation;
-import io.spine.server.command.given.CommandHandlerMethodTestEnv.InvalidHandlerNoParams;
-import io.spine.server.command.given.CommandHandlerMethodTestEnv.InvalidHandlerOneNotMsgParam;
-import io.spine.server.command.given.CommandHandlerMethodTestEnv.InvalidHandlerReturnsVoid;
-import io.spine.server.command.given.CommandHandlerMethodTestEnv.InvalidHandlerTooManyParams;
-import io.spine.server.command.given.CommandHandlerMethodTestEnv.InvalidHandlerTwoParamsFirstInvalid;
-import io.spine.server.command.given.CommandHandlerMethodTestEnv.InvalidHandlerTwoParamsSecondInvalid;
-import io.spine.server.command.given.CommandHandlerMethodTestEnv.ProcessManagerDoingNothing;
-import io.spine.server.command.given.CommandHandlerMethodTestEnv.RejectingAggregate;
-import io.spine.server.command.given.CommandHandlerMethodTestEnv.RejectingHandler;
-import io.spine.server.command.given.CommandHandlerMethodTestEnv.ValidHandlerButPrivate;
-import io.spine.server.command.given.CommandHandlerMethodTestEnv.ValidHandlerOneParam;
-import io.spine.server.command.given.CommandHandlerMethodTestEnv.ValidHandlerOneParamReturnsList;
-import io.spine.server.command.given.CommandHandlerMethodTestEnv.ValidHandlerTwoParams;
-import io.spine.server.command.given.CommandHandlerMethodTestEnv.ValidHandlerTwoParamsReturnsList;
+import io.spine.server.command.model.given.handler.HandlerReturnsEmpty;
+import io.spine.server.command.model.given.handler.HandlerReturnsEmptyList;
+import io.spine.server.command.model.given.handler.InvalidHandlerNoAnnotation;
+import io.spine.server.command.model.given.handler.InvalidHandlerNoParams;
+import io.spine.server.command.model.given.handler.InvalidHandlerOneNotMsgParam;
+import io.spine.server.command.model.given.handler.InvalidHandlerReturnsVoid;
+import io.spine.server.command.model.given.handler.InvalidHandlerTooManyParams;
+import io.spine.server.command.model.given.handler.InvalidHandlerTwoParamsFirstInvalid;
+import io.spine.server.command.model.given.handler.InvalidHandlerTwoParamsSecondInvalid;
+import io.spine.server.command.model.given.handler.ProcessManagerDoingNothing;
+import io.spine.server.command.model.given.handler.RejectingAggregate;
+import io.spine.server.command.model.given.handler.RejectingHandler;
+import io.spine.server.command.model.given.handler.ValidHandlerButPrivate;
+import io.spine.server.command.model.given.handler.ValidHandlerOneParam;
+import io.spine.server.command.model.given.handler.ValidHandlerOneParamReturnsList;
+import io.spine.server.command.model.given.handler.ValidHandlerTwoParams;
+import io.spine.server.command.model.given.handler.ValidHandlerTwoParamsReturnsList;
 import io.spine.server.model.HandlerMethodFailedException;
 import io.spine.server.procman.ProcessManager;
 import io.spine.test.reflect.ProjectId;
