+++ server/src/test/java/io/spine/server/command/model/given/handler/InvalidHandlerNoAnnotation.java
@@ -18,13 +18,18 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.command.model;
+package io.spine.server.command.model.given.handler;
 
-class CommandReactionMethodTest {
+import io.spine.core.CommandContext;
+import io.spine.test.reflect.command.RefCreateProject;
+import io.spine.test.reflect.event.RefProjectCreated;
 
+import static io.spine.server.model.given.Given.EventMessage.projectCreated;
 
-    /*
-     * Test environment.
-     */
-
+@SuppressWarnings("unused")
+// because the method is not annotated, which is the purpose of this test class.
+public class InvalidHandlerNoAnnotation extends TestCommandHandler {
+    public RefProjectCreated handleTest(RefCreateProject cmd, CommandContext context) {
+        return projectCreated(cmd.getProjectId());
+    }
 }
