+++ server/src/test/java/io/spine/server/command/model/given/handler/InvalidHandlerNoParams.java
@@ -18,13 +18,14 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.command.model;
+package io.spine.server.command.model.given.handler;
 
-class CommandReactionMethodTest {
+import io.spine.server.command.Assign;
+import io.spine.test.reflect.event.RefProjectCreated;
 
-
-    /*
-     * Test environment.
-     */
-
+public class InvalidHandlerNoParams extends TestCommandHandler {
+    @Assign
+    RefProjectCreated handleTest() {
+        return RefProjectCreated.getDefaultInstance();
+    }
 }
