+++ server/src/test/java/io/spine/server/command/model/given/handler/InvalidHandlerReturnsVoid.java
@@ -18,13 +18,14 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.command.model;
+package io.spine.server.command.model.given.handler;
 
-class CommandReactionMethodTest {
+import io.spine.core.CommandContext;
+import io.spine.server.command.Assign;
+import io.spine.test.reflect.command.RefCreateProject;
 
-
-    /*
-     * Test environment.
-     */
-
+public class InvalidHandlerReturnsVoid extends TestCommandHandler {
+    @Assign
+    void handleTest(RefCreateProject cmd, CommandContext context) {
+    }
 }
