+++ server/src/test/java/io/spine/server/command/model/given/handler/RejectingAggregate.java
@@ -18,13 +18,34 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.command.model;
+package io.spine.server.command.model.given.handler;
 
-class CommandReactionMethodTest {
+import com.google.protobuf.Empty;
+import io.spine.base.Identifier;
+import io.spine.server.aggregate.Aggregate;
+import io.spine.server.aggregate.Apply;
+import io.spine.server.command.Assign;
+import io.spine.server.entity.rejection.EntityAlreadyArchived;
+import io.spine.test.reflect.ProjectId;
+import io.spine.test.reflect.command.RefCreateProject;
+import io.spine.test.reflect.event.RefProjectCreated;
+import io.spine.validate.EmptyVBuilder;
 
+/**
+ * An aggregate which always rejects the passed command.
+ */
+public class RejectingAggregate extends Aggregate<ProjectId, Empty, EmptyVBuilder> {
+    public RejectingAggregate(ProjectId id) {
+        super(id);
+    }
 
-    /*
-     * Test environment.
-     */
+    @Assign
+    RefProjectCreated on(RefCreateProject cmd) throws EntityAlreadyArchived {
+        throw new EntityAlreadyArchived(Identifier.pack(cmd.getProjectId()));
+    }
 
+    @Apply
+    void event(RefProjectCreated evt) {
+        // Do nothing.
+    }
 }
