+++ server/src/test/java/io/spine/server/command/model/given/handler/TestCommandHandler.java
@@ -1,0 +1,55 @@
+/*
+ * Copyright 2018, TeamDev. All rights reserved.
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package io.spine.server.command.model.given.handler;
+
+import io.spine.server.BoundedContext;
+import io.spine.server.command.AbstractCommandHandler;
+
+import java.lang.reflect.Method;
+
+import static io.spine.util.Exceptions.newIllegalStateException;
+
+/**
+ * Abstract base for test environment command handlers.
+ */
+public abstract class TestCommandHandler extends AbstractCommandHandler {
+
+    private static final String HANDLER_METHOD_NAME = "handleTest";
+
+    protected TestCommandHandler() {
+        super(BoundedContext.newBuilder()
+                            .setMultitenant(true)
+                            .build()
+                            .getEventBus());
+    }
+
+    public Method getHandler() {
+        Method[] methods = getClass().getDeclaredMethods();
+        for (Method method : methods) {
+            if (method.getName()
+                      .equals(HANDLER_METHOD_NAME)) {
+                return method;
+            }
+        }
+        throw newIllegalStateException("No command handler method found: %s",
+                                       HANDLER_METHOD_NAME);
+    }
+}
