+++ server/src/test/java/io/spine/server/command/model/given/handler/ValidHandlerButPrivate.java
@@ -18,13 +18,19 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.command.model;
+package io.spine.server.command.model.given.handler;
 
-class CommandReactionMethodTest {
+import com.google.common.annotations.VisibleForTesting;
+import io.spine.server.command.Assign;
+import io.spine.test.reflect.command.RefCreateProject;
+import io.spine.test.reflect.event.RefProjectCreated;
 
+import static io.spine.server.model.given.Given.EventMessage.projectCreated;
 
-    /*
-     * Test environment.
-     */
-
+public class ValidHandlerButPrivate extends TestCommandHandler {
+    @Assign
+    @VisibleForTesting
+    public RefProjectCreated handleTest(RefCreateProject cmd) {
+        return projectCreated(cmd.getProjectId());
+    }
 }
